backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
celery==5.4.0
redis==5.2.1
psycopg[binary]==3.2.3
dj-database-url==2.3.0
gunicorn==23.0.0
